Changes to be committed: 	deleted:    IKONIC-TEST.zip 	modified:   app/Http/Controllers/MerchantController.php 	modified:   app/Http/Controllers/WebhookController.php 	modified:   app/Jobs/PayoutOrderJob.php 	modified:   app/Models/Merchant.php 	modified:   app/Models/Order.php 	modified:   app/Services/AffiliateService.php 	modified:   app/Services/ApiService.php 	modified:   app/Services/MerchantService.php 	modified:   app/Services/OrderService.php 	modified:   composer.json 	modified:   composer.lock 	modified:   database/factories/OrderFactory.php 	modified:   database/migrations/2022_05_13_220658_create_affiliates_table.php 	modified:   database/migrations/2022_05_16_143445_create_orders_table.php

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
-            // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            // Floating-point numbers are not able to precisely represent all decimal values, which can lead to unexpected results in calculations involving money or rates.
+            $table->decimal('commission_rate');
             $table->string('discount_code');
             $table->timestamps();
         });
